fix: rename positioner storybook file to fix docs
